- Installed heatmap.js: this required modifying the angular.json scripts object.
- Installed @types files for heatmap.js and leaflet.js but there was a problem...
heatmap.js had a node_modules subfolder with a different version of leaflet.js
which had to be deleted, and the @types/heatmap.js folder had the same.
Deleting this didn't work though, as the version was 0.7.33 instead of 1.5.5.
This types file didn't have any types for ILayers so I had to copy and paste
the relevant lines, which I'm sure is bad practise because you'll have to do the
same to get it to work...

- Modified home.page.html to change MapLayer to MapLayers

- added 4 icons to assets

- Several changes in home.page.ts:
	- Imported heatmap.js and a few other things from 'leaflet'
	- created some new variables in the constructor (without types - bad practice?
	- created four markers with different icons and added them to a layergroup
	- moved the creation of baselayer outside of options for readability
	- added a configuration for the heatmap
	- added some test data points
	- added a heatmap layer and set the data to the test data
	- updated the layersControl overlays from shapes to the new markers and heatmap
	- added the heatmaplayer to onMapReady so displays by default